+++ react/components/string-utils.ts
